Optimize.
For coercing to a string, simply perform a !== check against null. If
it succeeds, simply pass in the object, assuming that the JS
interpreter will be able to coerce faster than String() will. And if
the check fails simply pass in "null" directly.

And for functions that could go through the wrapper to avoid
complexity over the non-coercion of null, simply inline the null check
to save on the function call cost.